Fix a typo in the AdaptiveParameters sample problem.
`$self` should be `$ans`.  It seems that I changed the third parameter
of the checker from `$self` to the more customary `$ans`, but forgot to
change the usage in the method. Since that is used to skip the adaptive
parameter check when the equivalence message post filter is called, if
you submit an answer that is incorrect, but differs by a constant, and
then submit the correct answer you get the message that "your answer is
equivalent to the previous answer" even though the previous answer was
incorrect, but the current answer is correct.